+++ source/web/.eslintrc.json
@@ -11,19 +11,22 @@
     "sourceType": "module"
   },
   "rules": {
-    "indent": ["warn", 2, {"SwitchCase": 0}],
+    "indent": ["warn", 2, {"SwitchCase": 1}],
+    "camelcase": "off",
+    "eol-last": ["warn"],
     "quotes": [0],
-    "max-len": ["error", 120],
+    "max-len": ["off", 120],
     "no-unused-vars": "off",
+    "no-var": "off",
+    "object-curly-spacing": "warn",
+    "key-spacing": "warn",
     "@typescript-eslint/no-unused-vars": ["warn", { "varsIgnorePattern": "^(_|ui$|grok$|DG$)", "argsIgnorePattern": "^_"}],
     "require-jsdoc": "off",
     "valid-jsdoc": "warn",
     "spaced-comment": "off",
+    "space-before-function-paren": "off",
     "linebreak-style": "off",
-    "curly": [
-      "error",
-      "multi-or-nest"
-    ],
+    "curly": ["warn", "multi-or-nest", "consistent"],
     "brace-style": ["error", "1tbs", {"allowSingleLine": true}],
     "block-spacing": [2, "always"],
     "comma-dangle": [
