+++ source/web/package.json
@@ -6,9 +6,6 @@
   "sources": [
     "https://ajax.googleapis.com/ajax/libs/dojo/1.10.4/dojo/dojo.js.uncompressed.js"
   ],
-  "dependencies": {
-    "datagrok-api": "1.18.6"
-  },
   "devDependencies": {
     "@types/dojo": "^1.9.48",
     "@types/node": "^18.11.18",
@@ -19,7 +16,7 @@
     "eslint": "latest",
     "eslint-config-google": "latest",
     "source-map-loader": "^4.0.1",
-    "ts-loader": "^9.2.5",
+    "ts-loader": "^9.5.1",
     "typescript": "^4.8.4",
     "webpack": "^5.76.3",
     "webpack-cli": "^4.9.1"
