+++ source/web/src/types/index.ts
@@ -1,7 +1,3 @@
-import * as grok from 'datagrok-api/grok';
-import * as ui from 'datagrok-api/ui';
-import * as DG from 'datagrok-api/dg';
-
 import {ScilModuleType} from './scil';
 import {JSDraw2ModuleType} from './jsdraw2';
 
@@ -15,6 +11,5 @@ type JSDrawWindowType = {
 declare const window: JSDrawWindowType;
 
 export async function getJSDrawModules(): Promise<JSDrawWindowType> {
-  await grok.functions.call('JsDrawLite:ensureLoadJsDrawLite');
   return window;
 }
