+++ source/web/src/types/org.ts
@@ -1,0 +1,72 @@
+import {IAtom, IMol} from './jsdraw2';
+
+export const enum MonomerTypes {
+  BACKBONE = 'Backbone',
+  BRANCH = 'Branch',
+  TERMINAL = 'Terminal',
+}
+
+/** 'Backbone' | 'Branch' | 'Terminal' */
+export type MonomerType = `${MonomerTypes}`
+
+export const enum PolymerTypes {
+  RNA = 'RNA',
+  PEPTIDE = 'PEPTIDE',
+  CHEM = 'CHEM',
+  BLOB = 'BLOB',
+  G = 'G',
+}
+
+/** 'RNA' | 'PEPTIDE' | 'CHEM' | 'BLOB' | 'G' */
+export type PolymerType = `${PolymerTypes}`
+
+export const enum HelmTypes {
+  BASE = 'HELM_BASE',
+  SUGAR = 'HELM_SUGAR',
+  LINKER = 'HELM_LINKER',
+  AA = 'HELM_AA',
+  CHEM = 'HELM_CHEM',
+  BLOB = 'HELM_BLOB',
+  NUCLEOTIDE = 'HELM_NUCLETIDE',
+}
+
+/** 'HELM_BASE' | 'HELM_SUGAR' | 'HELM_LINKER' | 'HELM_AA' | 'HELM_CHEM' | 'HELM_BLOB' | 'HELM_NUCLETIDE' */
+export type HelmType = `${HelmTypes}`
+
+type IHelmTypes = Record<HelmTypes, OrgHelmType>;
+
+export type WebEditorRGroups = { [group: string]: string };
+
+export interface IMonomer {
+
+}
+
+export interface IOrgWebEditorMonomer {
+  /** symbol */ id: string;
+  /** name */ n?: string;
+  /** natural analog */ na?: string;
+  /* Pistoia.HELM deletes .type and .mt in Monomers.addOneMonomer() */
+  /** polymer type */type?: PolymerType;
+  /** monomer type */ mt?: MonomerType;
+  /** molfile */ m?: string;
+  /** substituents */ at: WebEditorRGroups;
+  /** number of substituents */ get rs(): number;
+
+  issmiles?: boolean;
+  smiles?: string;
+}
+
+export type IOrgWebEditor = {
+  HELM: typeof HelmTypes;
+  Interface: any;
+  isHelmNode(a: any): boolean;
+}
+
+export type OrgHelmType = {
+  webeditor: IOrgWebEditor;
+}
+
+export type OrgType = {
+  helm: OrgHelmType;
+}
+
