+++ source/web/webpack.config.js
@@ -17,19 +17,13 @@ module.exports = {
   },
   module: {
     rules: [
-      {test: /\.js$/, enforce: 'pre', use: ['source-map-loader'], exclude: /node_modules/},
-      {test: /\.ts(x?)$/, use: 'ts-loader', exclude: /node_modules/},
+      {test: /^.+\.js$/, enforce: 'pre', use: ['source-map-loader'], exclude: /node_modules/},
+      {test: /^.+\.ts(x?)$/, use: 'ts-loader', exclude: /node_modules/},
       {test: /\.css$/, use: ['style-loader', 'css-loader']},
     ],
   },
   plugins: [],
   devtool: mode !== 'production' ? 'inline-source-map' : 'source-map',
-  externals: {
-    'datagrok-api/dg': 'DG',
-    'datagrok-api/grok': 'grok',
-    'datagrok-api/ui': 'ui',
-    'wu': 'wu',
-  },
   output: {
     filename: '[name].js',
     library: packageName,
